up controllers, service

+++ services/contactService.js
@@ -1,32 +1,38 @@
 import { Contact } from "../models/contactModel.js";
 
-export const getContactsService = async () => { 
-   const contacts = await Contact.find()
+export const getContactsService = async (owner) => { 
+   const contacts = await Contact.find({ owner }, "-createdAt -updatedAt")
    return contacts;
 };
 
-export const deleteContactService = async (id) => {
-const removeContact = await Contact.findByIdAndDelete(id);
+export const deleteContactService = async (id, owner) => {
+const removeContact = await Contact.findByIdAndDelete({ _id: id,owner });
 return removeContact;
 };
 
-export const createContactService = async (contactData) => {
-    const newContact = await Contact.create(contactData);
+export const createContactService = async (contactData, owner) => {
+    const newContact = await Contact.create({...contactData, owner});
     return newContact;
 };
 
-export const updateContactService = async (id, contactData) => {
+export const updateContactService = async (id, contactData, owner) => {
     const upContact = await Contact.findByIdAndUpdate(
-        id,
+        {
+        _id: id,
+        owner,
+        },
         contactData,
         { new: true}
     );
         return upContact;
 };
 
-export const upStatusContactService = async (id, contactData) => {
+export const upStatusContactService = async (id, contactData, owner) => {
     const newStatus = await Contact.findByIdAndUpdate(
-    id,
+        {
+        _id: id,
+        owner,
+        },
     contactData,
      {new: true,},
 );
@@ -38,7 +44,7 @@ export const checkContactExistsService = async (filter) => {
 return contact
 };
 
-export const getContactsByIdService = async (id) => {
-    const contact = await Contact.findById(id);
+export const getContactsByIdService = async (id, owner) => {
+    const contact = await Contact.findById({ _id: id,owner });
 return contact
 };
